Rename to aws-elastic-search-encrypt-replication-group
