+++ internal/app/tfsec/rules/aws/elasticache/encrypt_replication_group_rule.go
@@ -1,4 +1,4 @@
-package elasticsearch
+package elasticache
 
 // generator-locked
 import (
