+++ pkg/externalscan/id_map.go
@@ -150,7 +150,6 @@ var idMap = map[string]string{
 	"aws-efs-enable-at-rest-encryption": "AVD-AWS-0037",
 	"aws-elastic-search-enforce-https": "AVD-AWS-0046",
 	"aws-elastic-search-enable-in-transit-encryption": "AVD-AWS-0043",
-	"aws-elastic-search-encrypt-replication-group": "AVD-AWS-0045",
 	"aws-elastic-search-use-secure-tls-policy": "AVD-AWS-0047",
 	"aws-elastic-search-enable-logging": "AVD-AWS-0044",
 	"aws-elastic-search-enable-domain-logging": "AVD-AWS-0042",
@@ -195,6 +194,7 @@ var idMap = map[string]string{
 	"aws-elbv2-alb-not-public": "AVD-AWS-0053",
 	"aws-elbv2-http-not-used": "AVD-AWS-0054",
 	"aws-codebuild-enable-encryption": "AVD-AWS-0018",
+	"aws-elasticache-encrypt-replication-group": "AVD-AWS-0045",
 	"aws-elasticache-enable-backup-retention": "AVD-AWS-0050",
 	"aws-elasticache-enable-in-transit-encryption": "AVD-AWS-0051",
 	"aws-elasticache-add-description-for-security-group": "AVD-AWS-0049",
